Replace DIRECTORY_SEPARATOR with '/' and force path strings to consistent use of '/'. Refs #3
Under Windows, DRIECTORY_SEPARATOR is set to '\', whereas it is '/' on *NIX. FileUtil and DataUtil, however, all try to normalize directory separators in paths to always use '/'. Attempting to use the DIRECTORY_SEPARATOR constant, especially when trying to pattern-match, will yield in unexpected results.

Replaced usage of DRIECTORY_SEPARATOR with '/', and also attempted to ensure that path strings were forced to always use '/' if it was not clear that the path string was coming in from a pass through DataUtil::formatForOS().